pull out msgs and state into separate files
