+++ erc20/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod contract;
+pub mod msg;
+pub mod state;
 
 #[cfg(test)]
 mod tests;
